[Refactor] I13: Split routes.go god-file into topic-grouped handler files
The previous internal/server/routes.go was 2828 lines and mixed route
registration with every O2-IMS handler body, callback validation, SSRF
IP range tables, and audit helpers. Route registration is now in
routes.go alongside only the cross-cutting wrappers (withPermission,
withPlatformAdmin, withTenantAccess, enforceTenantOwnershipInline,
resolveAdapter, setupRoutes, setupV1Routes). Handler methods keep their
*Server receiver and their exact semantics. Pure extraction, no
behavior change.

New files:
  internal/server/handlers_health.go         (health/readiness/metrics)
  internal/server/handlers_info.go           (root and /o2ims banners)
  internal/server/handlers_subscriptions.go  (Subscriptions CRUD)
  internal/server/handlers_resourcepools.go  (ResourcePools CRUD + validators)
  internal/server/handlers_resources.go      (Resources CRUD + URN validators)
  internal/server/handlers_resourcetypes.go  (ResourceTypes list/get)
  internal/server/handlers_deployment.go     (DeploymentManagers, O-Cloud)
  internal/server/handlers_tenant_quotas.go  (Tenant quota CRUD)
  internal/server/callback_validation.go     (ValidateCallback / ValidateCallbackHost)
  internal/server/audit.go                   (logAuditEvent)
  internal/server/privateip.go               (SSRF IP range helpers)

routes.go shrunk from 2828 to 412 lines. Build, lint (golangci-lint
./internal/server/... — zero issues), and the full go test ./... -short
suite all pass.

H9 (issue #482) was investigated in the same branch and determined to
be a larger architectural change than can be safely bundled with this
pure-extraction refactor without risking a false-security fix. See
issue #482 for the scoping comment and recommended dedicated-PR plan.

Closes #504
